back to eagle test

+++ deployment/deploy-docker.sh
@@ -18,7 +18,18 @@ fi
 # Pull the latest image (optional if you want to ensure it's the latest)
 echo "Pulling the latest image..."
 sudo docker pull dcronauer2025/cronauer-ceg3120:latest
+if [ $? -ne 0 ]; then
+  echo "Error: Failed to pull Docker image."
+  exit 1
+fi
 
 # Run the new container
 echo "Running the new container..."
 sudo docker run --rm -d --name $CONTAINER_NAME -p 80:4200 dcronauer2025/cronauer-ceg3120:latest
+if [ $? -ne 0 ]; then
+  echo "Error: Failed to run the Docker container."
+  exit 1
+fi
+
+echo "Container is running successfully!"
+
